refactor: export type WhitelistDocument = IWhitelist & Document;
