+++ packages/data/schema/system/src/model/system-settings/security_and_access/roles.schema.ts
@@ -1,6 +1,19 @@
 //@ts-ignore
 import mongoose, { Model, Schema } from 'mongoose';
-import { IRole, RoleDocument } from './interfaces';
+//import { IRole, RoleDocument } from './interfaces';
+import { ICreatedByAdmin } from '@isyss-cdm/interface';
+
+export interface IRole {
+	id?: string;
+	code: string;
+	displayName?: string;
+	description?: string;
+	createdAt?: Date;
+	updatedAt?: Date;
+	permissions?: mongoose.Types.ObjectId[]; // references to Permission docs
+	permissionsMeta?: Record<string, unknown>; // jsonb-like metadata
+	createdBy?: ICreatedByAdmin;
+}
 
 const roleSchema = new Schema(
 	{
@@ -17,8 +30,9 @@ const roleSchema = new Schema(
 
 roleSchema.index({ code: 1 }, { unique: true });
 
+export type RoleDocument = IRole & Document;
 export const Role: Model<RoleDocument> =
 	(mongoose.models.Role as Model<RoleDocument>) || mongoose.model<RoleDocument>('Role', roleSchema);
 
-export { IRole, RoleDocument };
+//export { IRole, RoleDocument };
 export default Role;
