+++ packages/data/schema/system/src/model/system-settings/security_and_access/security_config.schema.ts
@@ -1,6 +1,23 @@
 //@ts-ignore
 import mongoose, { Model, Schema } from 'mongoose';
-import { ISecurityConfig, SecurityConfigDocument } from './interfaces';
+//import { ISecurityConfig, SecurityConfigDocument } from './interfaces';
+import { ICreatedByAdmin } from '@isyss-cdm/interface';
+
+export interface ISecurityConfig {
+	id?: string;
+	env?: string;
+	rateLimiting?: Record<string, unknown>;
+	mfa?: Record<string, unknown>;
+	passwordPolicy?: Record<string, unknown>;
+	rbac?: Record<string, unknown>;
+	ipControls?: Record<string, unknown>;
+	tokenPolicy?: Record<string, unknown>;
+	auditLogging?: Record<string, unknown>;
+	createdAt?: Date;
+	updatedAt?: Date;
+	createdBy?: ICreatedByAdmin;
+}
+
 
 const securityConfigSchema = new Schema(
 	{
@@ -20,10 +37,10 @@ const securityConfigSchema = new Schema(
 
 securityConfigSchema.index({ env: 1 });
 securityConfigSchema.index({ env: 1 }, { unique: true });
-
+export type SecurityConfigDocument = ISecurityConfig & Document;
 export const SecurityConfig: Model<SecurityConfigDocument> =
 	(mongoose.models.SecurityConfig as Model<SecurityConfigDocument>) ||
 	mongoose.model<SecurityConfigDocument>('SecurityConfig', securityConfigSchema);
 
-export { ISecurityConfig, SecurityConfigDocument };
+//export { ISecurityConfig, SecurityConfigDocument };
 export default SecurityConfig;
